requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
factory-boy==3.3.0
